refactor(cli): move Windows paste capture into session::paste
  Extract the paste subsystem — the Windows console FFI, the cross-platform chip model, and the rustyline handlers — out of the core input
  module and into a new session::paste module. The change is a verbatim move: input.rs drops 454 lines and gains only a four-item
  pub(super) import, with no behavior change. This addresses the reviewer's concern that the fix was a very large change to a core module
